Refactor entity classes to use Long for ID fields; update related services, controllers, and database initialization scripts for consistency. Enhance restaurant table order view with additional fields and update Docker Compose for service health checks.

+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/employee/dto/RestaurantDishView.java
@@ -8,11 +8,11 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class RestaurantDishView {
-    private Integer id;
+    private Long id;
     private String name;
     private String categoryName;
     private String description;
     private BigDecimal price;
     private Boolean available;
-    private Integer restaurantId;
+    private Long restaurantId;
 }
