+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/shared/entity/OrderItemsEntity.java
@@ -4,7 +4,7 @@ import jakarta.persistence.Table;
 
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -53,5 +53,5 @@ public class OrderItemsEntity {
      * Total de la orden.
      */
     @Column(name = "total", nullable = false, precision = 10, scale = 2)
-    private Double total;
+    private BigDecimal total;
 }
